- nvrhi: fix buffer alignment to 4

+++ core/materialsystem1/Renderers/NVRHI/NVRHIBuffer.h
@@ -17,7 +17,7 @@ public:
 	~CNVRHIBuffer();
 	CNVRHIBuffer(const BufferInfo& bufferInfo, int bufferUsageFlags, const char* label = nullptr);
 
-	int			GetSize() const { return m_bufSize; }
+	int64		GetSize() const { return m_bufSize; }
 
 	void		Update(const void* data, int64 size, int64 offset);
 	MapFuture	Lock(int lockOfs, int sizeToLock, int flags);
@@ -36,7 +36,7 @@ private:
 	EqString	m_dbgName;
 
 	nvrhi::BufferHandle	m_rhiBuffer{ nullptr };
-	int			m_bufSize{ 0 };
+	int64		m_bufSize{ 0 };
 	int			m_usageFlags{ 0 };
 	bool		m_isLocked{ false };
 	bool		m_needsTrackingState{ true };
